Supports for the aud claim as string or array
This was created based on the following PR:
https://github.com/dgrijalva/jwt-go/pull/286

As per the JWT spec, the aud claim field can be either a single string value or an array of strings.

jwt-go would completely drop array values as the StandardClaims struct's Audience field is a string value and the value is dropped upon deserialization.